modularizing activity select
