+++ package.json
@@ -32,6 +32,5 @@
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject"
-   
   }
 }
