+++ src/components/Events/WriteActivity.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import { fetchCategories, fetchActivities } from '../../actions/Activities';
+import { updateNewEventState } from '../../actions/New-Event';
+import moment from 'moment';
+
+export default class WriteActivity extends React.Component {
+  render(){
+    return (
+      <div>write</div>
+    );
+  }
+}
